Tasks 7 and 17
Added error handling and prevention to courses screen to prevent issues if data is missing from the database.
Added an accompanying dictionary to the study plan file to save whether a class has been completed or not. Added handling to the save and load functions to account for the new object. Added course_taken button handler to change background color of the button clicked to indicate that a course has been completed.

Tested all additions and basic searching and study plan handlign to verify no other errors introduced.

Task 8, multi-file study plans, was deferred to Alla.